requirements.txt
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.14.0
yfinance>=0.2.0
scipy>=1.10.0
requests>=2.31.0
gspread>=5.10.0
google-auth>=2.22.0
